+++ Pandora/Paradigm/Structure/Concrete.hs
@@ -1,4 +1,5 @@
 module Pandora.Paradigm.Structure.Concrete (module Exports) where
 
+import Pandora.Paradigm.Structure.Concrete.Binary as Exports
 import Pandora.Paradigm.Structure.Concrete.Graph as Exports
 import Pandora.Paradigm.Structure.Concrete.Stack as Exports